Fix unknown typename on mac

+++ include/pyro/pyrostring.h
@@ -3,6 +3,6 @@
 
 #include <string>
 namespace Pyro {
-    std::string nf(uint number, uint digits);
+    std::string nf(uint64_t number, uint64_t digits);
 }
 #endif
